move the regression test into its own file
node-http2.test.js has a pre-existing borderline-timeout test
(maxSessionMemory, ~100s on debug+ASAN against a 150s limit) that flakes
independently of this change. Moving the new test into its own file keeps
the fail-before/pass-after proof deterministic.
